requirements.txt
requests
dnspython
pyOpenSSL
